requirements.txt
annotated-types==0.7.0
markdown-it-py==3.0.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
rich==13.7.1
beautifulsoup4==4.12.3
matplotlib==3.9.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.5.1
seaborn==0.13.2
setuptools==72.1.0
tqdm==4.66.5
yt-dlp==2024.8.6
ffmpeg==1.4
selenium==4.23.1
webdriver-manager==4.0.2
pandas==2.2.2
torcheval==0.0.7
fastapi==0.111.0
fastapi-cli==0.0.4
pydantic==2.7.4
pydantic_core==2.18.4
PyMySQL==1.1.1
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
pytest-twisted==1.14.2
python-dotenv==1.0.1
uvicorn==0.30.1
watchfiles==0.22.0
aiofiles==24.1.0
requests==2.32.3
pydantic-settings
moviepy
librosa
torchmetrics
transformers
munch
webrtcvad
nltk
spacy
textattack
twisted
aiobotocore
cryptography
bcrypt